+++ midap/segmentation/unet_segmentator.py
@@ -19,7 +19,7 @@ class UNetSegmentation(SegmentationPredictor):
     A class that performs the image segmentation of the cells using a UNet
     """
 
-    supported_setups = ["Family_Machine"]
+    supported_setups = ["Family_Machine", "Mother_Machine"]
 
     def __init__(self, *args, **kwargs):
         """