+++ CHANGELOG.md
@@ -19,12 +19,21 @@ Efficiency:
 Fix:
 - ...
 
-## [0.3.12]
+## [0.3.13]
+
+TBD
 
 Feature
+- Mother machine UNet training
+- Mother machine UNet support
+
+
+
+## [0.3.12]
 
 2023-08-17
 
+Feature
 - New advanced options in GUI
 - Brightness thresholding and notebook
 - Support for single image runs
@@ -34,10 +43,9 @@ Feature
 
 ## [0.3.11]
 
-Feature
-
 2023-05-10
 
+Feature
 - Added fine-tuned omnipose models and upped the model weights version in the json.
 - Update env file accordingly